fix: 타입분리, export 분리 등등 수정
