+++ src/api/member_check_api.ts
@@ -1,20 +1,10 @@
 import api from "../utils/api";
 import { useTokenStore } from "../stores/tokenStore";
+import { MemberCheckResponse } from "../types/member_check";
 
-export interface MemberCheckResponse {
-  code: string;          
-  data: {
-    nickname:      string;
-    oauthId:       number;
-    oauthProvider: string;
-  };
-  message: string;
 
-  accessToken?:  string;
-  refreshToken?: string;
-}
 
-export const memberCheck = async (
+const memberCheck = async (
   accessToken: string
 ): Promise<MemberCheckResponse> => {
   const { data } = await api.post<MemberCheckResponse>(
@@ -29,3 +19,5 @@ export const memberCheck = async (
 
   return data;
 };
+
+export { memberCheck };
