+++ src/types/member_check.ts
@@ -1,0 +1,12 @@
+export interface MemberCheckResponse {
+  code: string;          
+  data: {
+    nickname:      string;
+    oauthId:       number;
+    oauthProvider: string;
+  };
+  message: string;
+
+  accessToken?:  string;
+  refreshToken?: string;
+}
