+++ src/types/refresh.ts
@@ -1,0 +1,8 @@
+export interface RefreshPayload {
+  refreshToken: string;
+}
+
+export interface RefreshResponse {
+  accessToken:  string;
+  refreshToken: string;
+}
